Commented out Update Document section for swift. Updated DeleteDocument section. Cleaned up unbookmark code logic

+++ ios/TravelSample/TravelSample/Presenter/HotelPresenter.swift
@@ -106,14 +106,7 @@ extension HotelPresenter {
                 handler(TravelSampleError.DocumentFetchException)
                 return
             }
-            
-            // Get current list of bookmarked hotels.
-            guard let arrOfCurrentIds:[String] = document.array(forKey: "hotels")?.toArray().flatMap({ return $0 as? String }) else {
-                handler(TravelSampleError.DocumentFetchException)
-                return
-            }
-            
-            
+        
             // Get the Ids of all hotels that need to be unbookmarked from the hotels array
             let idsToRemove:[String] = hotels.map({ (dict)  in
                 if let idVal = dict["id"] as? String {
@@ -122,27 +115,32 @@ extension HotelPresenter {
                 return ""
             })
             
+            
+            // Get current list of bookmarked hotels.
+            guard let arrOfCurrentIds:[String] = document.array(forKey: "hotels")?.toArray().flatMap({ return $0 as? String }) else {
+                handler(TravelSampleError.DocumentFetchException)
+                return
+            }
+            
+            
             // Convert to Set for easy set operations
             let setOfCurrentBookmarkedIds:Set<String> = Set(arrOfCurrentIds)
             
             // Get the delta of the new list and current Id list to identify the hotels that are not yet bookmarked
             let IdToRemain = Array(setOfCurrentBookmarkedIds.subtracting(idsToRemove))
-            
-                // perform batch update
-                try db.inBatch {
-                    // Update the bookmarked Id list
-                    document.setArray(ArrayObject.init(array: IdToRemain), forKey: "hotels")
-                    // Save updated version of bookmarkedhotels document
-                    try db.save(document)
+         
+            // Update the bookmarked Id list
+            document.setArray(ArrayObject.init(array: IdToRemain), forKey: "hotels")
+            // Save updated version of bookmarkedhotels document
+            try db.save(document)
                     
-                    // Remove unbookmarked hotel documents
-                    for idOfDocToRemove in idsToRemove {
-                        if let doc = db.getDocument(idOfDocToRemove) {
-                            try db.delete(doc)
-                        }
-                    }
-                    
+            // Remove unbookmarked hotel documents
+            for idOfDocToRemove in idsToRemove {
+                if let doc = db.getDocument(idOfDocToRemove) {
+                    try db.delete(doc)
                 }
+            }
+                
             handler(nil)
 
         }
